crawler/requirements.txt
annotated-types==0.7.0
attrs==24.3.0
autoflake==2.3.1
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
deepsearch-glm==1.0.0
dill==0.3.9
docling==2.15.1
docling-core==2.14.0
docling-ibm-models==3.1.2
docling-parse==3.0.0
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.16.1
filetype==1.2.0
fsspec==2024.12.0
huggingface-hub==0.27.1
idna==3.10
imageio==2.36.1
Jinja2==3.1.5
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
lxml==5.3.0
markdown-it-py==3.0.0
marko==2.1.2
MarkupSafe==3.0.2
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.17
mysql-connector==2.2.9
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.1
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pyclipper==1.3.0.post6
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
pypdfium2==4.30.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
Rtree==1.3.0
safetensors==0.5.2
scikit-image==0.25.0
scipy==1.15.1
semchunk==2.2.2
setuptools==75.8.0
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
soupsieve==2.6
sympy==1.13.1
tabulate==0.9.0
tifffile==2025.1.10
tokenizers==0.21.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.48.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wheel==0.45.1
XlsxWriter==3.2.0
